+++ bzt/modules/dotnet/executors.py
@@ -1,5 +1,5 @@
 """
-Copyright 2017 BlazeMeter Inc.
+Copyright 2018 BlazeMeter Inc.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,12 +14,12 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 import os
-from subprocess import check_output, CalledProcessError, STDOUT
 
-from bzt import TaurusConfigError, ToolError
+from bzt import TaurusConfigError
 from bzt.modules import SubprocessedExecutor
 from bzt.engine import HavingInstallableTools
-from bzt.utils import get_full_path, is_windows, RequiredTool
+from bzt.utils import get_full_path, is_windows
+from .tools import Mono, Dotnet
 
 
 class NUnitExecutor(SubprocessedExecutor, HavingInstallableTools):
@@ -66,18 +66,38 @@ class NUnitExecutor(SubprocessedExecutor, HavingInstallableTools):
         self._start_subprocess(cmdline)
 
 
-class Mono(RequiredTool):
-    def __init__(self, tool_path, download_link, parent_logger):
-        super(Mono, self).__init__("Mono", tool_path, download_link)
-        self.log = parent_logger.getChild(self.__class__.__name__)
+class XUnitExecutor(SubprocessedExecutor, HavingInstallableTools):
+    def __init__(self):
+        super(XUnitExecutor, self).__init__()
+        self.runner_dir = os.path.join(get_full_path(__file__, step_up=3), "resources", "XUnitRunner")
+        self.runner_executable = os.path.join(self.runner_dir, "XUnitRunner.dll")
+        self.dotnet = Dotnet("dotnet", "", self.log)
 
-    def check_if_installed(self):
-        try:
-            output = check_output([self.tool_path, '--version'], stderr=STDOUT)
-            self.log.debug("%s output: %s", self.tool_name, output)
-            return True
-        except (CalledProcessError, OSError):
-            return False
+    def install_required_tools(self):
+        if not is_windows():
+            if not self.dotnet.check_if_installed():
+                self.dotnet.install()
 
-    def install(self):
-        raise ToolError("%s is not operable or not available. Consider installing it" % self.tool_name)
+    def prepare(self):
+        super(XUnitExecutor, self).prepare()
+        self.script = get_full_path(self.get_script_path())
+        if not self.script:
+            raise TaurusConfigError("Script not passed to runner %s" % self)
+
+        self.install_required_tools()
+        self.reporting_setup(suffix=".ldjson")
+
+    def startup(self):
+        cmdline = [self.dotnet.tool_path]
+
+        cmdline += [self.runner_executable,
+                    "--target=" + self.script,
+                    "--report-file=" + self.report_file]
+
+        load = self.get_load()
+        if load.iterations:
+            cmdline += ['--iterations=' + str(load.iterations)]
+        if load.hold:
+            cmdline += ['--duration=' + str(int(load.hold))]
+
+        self._start_subprocess(cmdline)
